Ladybird/GTK: Separate out internal WebView API into WebViewPrivate.h
